Move config.py into numba_dpex/core and fix circular imports
