+++ numba_dpex/core/passes/passes.py
@@ -13,7 +13,7 @@ from numba.core.compiler_machinery import (
 from numba.core.ir_utils import remove_dels
 from numba.core.typed_passes import NativeLowering
 
-from numba_dpex import config
+from numba_dpex.core import config
 
 
 @register_pass(mutates_CFG=True, analysis_only=False)
