+++ numba_dpex/spirv_generator.py
@@ -8,7 +8,7 @@ import os
 import tempfile
 from subprocess import STDOUT, CalledProcessError, check_output
 
-from numba_dpex import config
+from numba_dpex.core import config
 from numba_dpex.core.exceptions import InternalError
 from numba_dpex.core.targets.kernel_target import LLVM_SPIRV_ARGS
