+++ numba_dpex/tests/__init__.py
@@ -3,4 +3,4 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import numba_dpex
-from numba_dpex import config
+from numba_dpex.core import config
