+++ numba_dpex/tests/_helper.py
@@ -12,7 +12,8 @@ import dpctl
 import dpnp
 import pytest
 
-from numba_dpex import config, dpjit, numba_sem_version
+from numba_dpex.core import config
+from numba_dpex import dpjit, numba_sem_version
 
 
 @cache
