+++ numba_dpex/utils/array_utils.py
@@ -8,7 +8,7 @@ import dpctl
 import dpctl.memory as dpctl_mem
 import numpy as np
 
-from numba_dpex import config
+from numba_dpex.core import config
 
 supported_numpy_dtype = [
     np.int32,
